Restored WORKSPACE
Restored the original WORKSPACE since the bzlmod work is still marked experimental.
Fixed the module to work along side the original build.
Update the Download SDK to 1.20

Even though there is less text, the clarity of the work in the MODULE is at best vague. Several iterations had to occur in order to get the syntax and order correct to build with the module and the non module work.